Reorganize evaluator visitor code
