+++ excellent/base.go
@@ -1,0 +1,59 @@
+package excellent
+
+import (
+	"strings"
+
+	"github.com/nyaruka/goflow/excellent/gen"
+
+	"github.com/antlr/antlr4/runtime/Go/antlr"
+)
+
+// BaseVisitor is the base of visitors
+type BaseVisitor struct {
+	gen.BaseExcellent2Visitor
+}
+
+// VisitExpression parses and visits the given expression with the given visitor
+func VisitExpression(expression string, visitor antlr.ParseTreeVisitor) (interface{}, error) {
+	errListener := NewErrorListener(expression)
+
+	input := antlr.NewInputStream(expression)
+	lexer := gen.NewExcellent2Lexer(input)
+	stream := antlr.NewCommonTokenStream(lexer, 0)
+	p := gen.NewExcellent2Parser(stream)
+	p.RemoveErrorListeners()
+	p.AddErrorListener(errListener)
+	tree := p.Parse()
+
+	// if we ran into errors parsing, return the first one
+	if len(errListener.Errors()) > 0 {
+		return nil, errListener.Errors()[0]
+	}
+
+	return visitor.Visit(tree), nil
+}
+
+// VisitTemplate scans the given template and calls the callback for each token encountered
+func VisitTemplate(template string, allowedTopLevels []string, callback func(XTokenType, string) error) error {
+	scanner := NewXScanner(strings.NewReader(template), allowedTopLevels)
+	errors := NewTemplateErrors()
+
+	for tokenType, token := scanner.Scan(); tokenType != EOF; tokenType, token = scanner.Scan() {
+		err := callback(tokenType, token)
+		if err != nil {
+			var repr string
+			if tokenType == IDENTIFIER {
+				repr = "@" + token
+			} else {
+				repr = "@(" + token + ")"
+			}
+
+			errors.Add(repr, err.Error())
+		}
+	}
+
+	if errors.HasErrors() {
+		return errors
+	}
+	return nil
+}
